backend/requirements.txt
# backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication (commented out for now)
fastapi-users[sqlalchemy]==12.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI & ML
google-generativeai==0.3.2
chromadb==0.4.18
numpy==1.24.3
requests==2.31.0

# File processing
PyPDF2==3.0.1
python-docx==1.1.0

# Export functionality
reportlab==4.0.7
openpyxl==3.1.2
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0